Migrate Task and ProcessFile to app
