+++ gridcapa-task-manager-api/src/test/java/com/farao_community/farao/gridcapa/task_manager/api/ProcessFileDtoTest.java
@@ -7,9 +7,6 @@
 package com.farao_community.farao.gridcapa.task_manager.api;
 
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-
-import java.time.LocalDateTime;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -18,23 +15,6 @@ import static org.junit.jupiter.api.Assertions.assertNull;
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
  */
 class ProcessFileDtoTest {
-
-    @Test
-    void testConstructorFromEntity() {
-        Task task = Mockito.mock(Task.class);
-        ProcessFile processFile = new ProcessFile(task, "CGM");
-        processFile.setFileUrl("cgm-url");
-        processFile.setFilename("cgm-name");
-        processFile.setLastModificationDate(LocalDateTime.parse("2021-10-11T10:18"));
-        processFile.setProcessFileStatus(ProcessFileStatus.VALIDATED);
-
-        ProcessFileDto processFileDto = ProcessFileDto.fromEntity(processFile);
-        assertEquals("CGM", processFileDto.getFileType());
-        assertEquals("cgm-url", processFileDto.getFileUrl());
-        assertEquals("cgm-name", processFileDto.getFilename());
-        assertEquals(LocalDateTime.parse("2021-10-11T10:18"), processFileDto.getLastModificationDate());
-        assertEquals(ProcessFileStatus.VALIDATED, processFileDto.getProcessFileStatus());
-    }
 
     @Test
     void testConstructorEmptyProcessFile() {
